[ADF-2974] Moved mock inside testing file for buttons menu component
