--- lib/core/mock/buttons-menu.mock.ts
@@ -1,47 +1,0 @@
-/*!
- * @license
- * Copyright 2016 Alfresco Software, Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'adf-custom-container',
-    template: `
-        <adf-buttons-action-menu>
-            <button mat-menu-item (click)="assignValue()">
-                <mat-icon>settings</mat-icon><span> Button </span>
-            </button>
-        </adf-buttons-action-menu>
-    `
-})
-export class CustomContainerComponent {
-
-    value: number;
-
-    assignValue() {
-        this.value = 1;
-    }
-}
-
-@Component({
-    selector: 'adf-custom-empty-container',
-    template: `
-        <adf-buttons-action-menu>
-        </adf-buttons-action-menu>
-    `
-})
-export class CustomEmptyContainerComponent {
-}
