+++ lib/core/mock/public-api.ts
@@ -28,7 +28,6 @@ export * from './renditionsService.mock';
 export * from './search.service.mock';
 export * from './translation.service.mock';
 export * from './alfresco-api.service.mock';
-export * from './buttons-menu.mock';
 
 export * from './form/form.component.mock';
 export * from './form/formDefinition.mock';
